Notifications update
"return camera(id, model)" doesn't work as intended. if you look at the
URL it keeps the /toggleOn?id= which shouldn't be the case because then
the attributes in /camera don't update. should be return
"redirect:/hub/camera"; but not sure how to get the parameters in there.
this may be where flash attributes come in

been trying a lot of different things to get notifications working but
havent fully grasped it yet. I'm trying to avoid the "chat" method
recommended because its confusing as I don't really know spring. Have to
push to master because i realized i broke something untested.